\name{sigma2tools-package}
\alias{sigma2tools-package}
\docType{package}
\title{
Test of hypothesis about sigma2
}
\description{
This package contents a set of functions to investigate
with sigma2 parameter
}
\details{
\tabular{ll}{
Package: \tab sigma2tools\cr
Type: \tab Package\cr
Encoding: \tab latin1\cr
Version: \tab 1.2.5\cr
Date: \tab 2006-09-10\cr
Depends: \tab R, stats, utils, tools\cr
License: \tab GPL-2\cr
}

Index:
\preformatted{
codev                   Calcola la codevianza
cv2                     Calcola il coefficiente di variazione
                        campionario
midrange                Calcola il midrange
popstderror             Calcola l'errore standard della popolazione
range2                  Calcola il campo di variazione
sigma                   Calcola lo scarto quadratico medio
sigma2.test             Test di ipotesi per la varianza
sigma2m                 Calcola la matrice di varianza e covarianza non
                        corretta
sigma2tools-package     Test of hypothesis about sigma2
ssdev                   Calcola la devianza
stderror                Calcola l'errore standard campionario
}
}
\author{
Fabio Frascati

Maintainer: Fabio Frascati <fabiofrascati@yahoo.it>
}
\keyword{ package }
